requirements.txt
awsebcli==3.7.6
backports.ssl-match-hostname==3.5.0.1
blessed==1.9.5
botocore==1.4.27
cement==2.8.2
colorama==0.3.7
Django==1.9.7
django-filter==0.13.0
djangorestframework==3.3.3
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.12
jmespath==0.9.0
pathspec==0.3.4
psycopg2==2.6.1
python-dateutil==2.5.3
PyYAML==3.11
requests==2.9.1
semantic-version==2.5.0
six==1.10.0
texttable==0.8.4
wcwidth==0.1.6
websocket-client==0.37.0
